requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2019.11.28
chardet==3.0.4
chromedriver==2.24.1
chromedriver-binary==84.0.4147.30.0
colorama==0.4.3
discord.py==1.3.4
gspread==3.3.0
httplib2==0.17.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
numpy==1.19.1
oauth2client==4.1.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
pylint==2.4.4
requests==2.23.0
rsa==4.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
